회원가입 , 로그인 , 차량 목록 프론트 API 구현

+++ src/main/java/com/carrentall/backend/vehicle/dto/VehicleResponse.java
@@ -20,7 +20,7 @@ public class VehicleResponse {
     private Long hourlyRate;
     private Long dailyRate;
 
-    public VehicleResponse(Long id, String manufacturer, String modelName, String vehicleNumber, RentalType rentalType, FuelType fuelType, VehicleStatus status, Integer hourlyRate, Integer dailyRate) {
+    public VehicleResponse(Long id, String manufacturer, String modelName, String vehicleNumber, RentalType rentalType, FuelType fuelType, VehicleStatus status, Long hourlyRate, Long dailyRate) {
         this.id = id;
         this.manufacturer = manufacturer;
         this.modelName = modelName;
